re-organise the test harness code to improve the imports
