+++ test-harness/py_harness/test_cases/base.py
@@ -1,7 +1,3 @@
-import logging
-import time
-from typing import List
-from host import LinkedCloneHost
 from .test_case import TestCase, registered_test_cases
 from .shared_runtime_tests import *
 from .test_case import TestReport
