+++ test-harness/py_harness/test_cases/shared_runtime_tests.py
@@ -1,7 +1,7 @@
 import subprocess
-import harness_settings
-from host import LinkedCloneHost
-from guest_control import ssh_command
+from py_harness.harness_settings import test_case_location, base_ssh_key
+from py_harness.config.host import LinkedCloneHost
+from py_harness.config.guest_control import ssh_command
 from .test_case import TestReport
 from typing import List
 import inspect
@@ -12,12 +12,12 @@ def check_internet_connectivity(
         linked_clone_hosts: List[LinkedCloneHost],
         guest_names: List[str],
 ) -> TestReport:
-    test_case = f"{harness_settings.test_case_location}/{test_case_name}"
+    test_case = f"{test_case_location}/{test_case_name}"
     # for each guest supplied (by names in the yaml file), try to curl google as a network test
     results = []
     for guest_name in guest_names:
         curl_result: subprocess.CompletedProcess = ssh_command(f"cd {test_case} && kvm-compose exec {guest_name} shell-command curl google.com",
-                            harness_settings.base_ssh_key,
+                            base_ssh_key,
                             linked_clone_hosts[0].hostname,  # first host will be main
                             )
         results.append([guest_name, curl_result])
