+++ test-harness/py_harness/test_cases/test_case.py
@@ -1,11 +1,11 @@
 import time
 import logging
-import harness_settings
+from harness_settings import test_case_location, base_ssh_key
 from datetime import datetime
 from abc import ABC, abstractmethod
 from typing import List, Optional
-from host import LinkedCloneHost
-from guest_control import ssh_command
+from py_harness.config.host import LinkedCloneHost
+from py_harness.config.guest_control import ssh_command
 
 # register the test cases here in the files they are written
 registered_test_cases = []
@@ -108,12 +108,12 @@ def deploy_test_case(example_name: str, linked_clone_hosts: List[LinkedCloneHost
     # the tests we want to use exist in the test harness folder, which will have the relevant kvm-compose.yaml files
 
     # get the location of the test, where the kvm-compose.yaml exists
-    test_case = f"{harness_settings.test_case_location}/{example_name}"
+    test_case = f"{test_case_location}/{example_name}"
     # make sure the testbed-server is running as a daemon on all hosts
     for host in linked_clone_hosts:
         logging.info("Making sure the testbed server is running on the hosts")
         ssh_command(f"sudo systemctl restart testbedos-server.service",
-                    harness_settings.base_ssh_key,
+                    base_ssh_key,
                     host.hostname,
                     )
     # need to give the server a second to start before it can accept commands
@@ -121,7 +121,7 @@ def deploy_test_case(example_name: str, linked_clone_hosts: List[LinkedCloneHost
     # on the main testbed, run the up command for the test case in the correct location
     logging.info(f"Run up command on '{test_case}'")
     up_result = ssh_command(f"cd {test_case} && kvm-compose up",
-                harness_settings.base_ssh_key,
+                base_ssh_key,
                 linked_clone_hosts[0].hostname,  # first host will be main
                 )
 
@@ -133,11 +133,11 @@ def destroy_test_case(example_name: str, linked_clone_hosts: List[LinkedCloneHos
     # re-usable function to destroy a test harness test case on the testbed ...
 
     # get the location of the test, where the kvm-compose.yaml exists
-    test_case = f"{harness_settings.test_case_location}/{example_name}"
+    test_case = f"{test_case_location}/{example_name}"
     # on the main testbed, run the down command for the test case in the correct location
     logging.info(f"Run down command on '{test_case}'")
     down_result = ssh_command(f"cd {test_case} && kvm-compose down",
-                            harness_settings.base_ssh_key,
+                            base_ssh_key,
                             linked_clone_hosts[0].hostname,  # first host will be main
                             )
 
@@ -149,12 +149,12 @@ def clear_artefacts(example_name: str, linked_clone_hosts: List[LinkedCloneHost]
     # re-usable function to remove the test case artefacts (vm images etc) once the test case has been destroyed
 
     # get the location of the test, where the kvm-compose.yaml exists
-    test_case = f"{harness_settings.test_case_location}/{example_name}"
+    test_case = f"{test_case_location}/{example_name}"
 
     # on the main testbed, run the clear artefacts command for the test case in the correct location
     logging.info(f"Run clear-artefacts command on '{test_case}'")
     clear_artefacts_result = ssh_command(f"cd {test_case} && kvm-compose clear-artefacts",
-                              harness_settings.base_ssh_key,
+                              base_ssh_key,
                               linked_clone_hosts[0].hostname,  # first host will be main
                               )
 
@@ -165,7 +165,7 @@ def clear_artefacts(example_name: str, linked_clone_hosts: List[LinkedCloneHost]
     # finally also remove the state json, as this would prevent future up commands from working
     logging.info(f"Remove state json for '{test_case}'")
     rm_state_json_result = ssh_command(f"cd {test_case} && rm {example_name}-state.json",
-                              harness_settings.base_ssh_key,
+                              base_ssh_key,
                               linked_clone_hosts[0].hostname,  # first host will be main
                               )
